refactor: 로또 번호 생성기 유틸클래스로 변경
