+++ src/test/java/lotto/domain/util/LottoNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -14,11 +14,10 @@ class LottoNumberGeneratorTest {
     @Test
     @DisplayName("로또 번호는 1~45까지의 범위의 수만 생성해야 합니다.")
     void test01() {
-        LottoNumberGenerator lottoNumberGenerator = new LottoNumberGenerator();
         Set<Integer> numberSet = new HashSet<>();
 
         IntStream.range(0, 10000).forEach(n -> {
-            numberSet.add(lottoNumberGenerator.generate());
+            numberSet.add(LottoNumberGenerator.generate());
         });
 
         assertThatNoException().isThrownBy(() -> validateSet(numberSet));
